Use the non-throwing variant of `std::filesystem::exists()` in `rocky::URI::read()`.
There appears to be a bug in the MSVC STL implementation of
`std::filesystem::exists()` in which the mechanism used to
determine whether or not a given file is found (that is, the
`__std_is_file_not_found()` function, although that is not a part
of the C++ Standard) does not consider every possible relevant
Win32 error code which could be returned by
`GetFileAttributesExW()`. (This issue has been documented in a pull
request at https://github.com/microsoft/STL/pull/5381.)

This commit implements a work around for that issue by modifying
the `rocky::URI::read()` function to use the non-throwing variant
of `std::filesystem::exists()` in order to determine whether the
URI returned from `rocky::URI::full()` represents an existing file
in the local filesystem. This variant of the function will return
`false` if the specified `std::filesystem::path` does not represent
an existing object in the filesystem, just like the throwing
variant. In addition, it will also return `false` if an error
prevented the function from determining whether or not the file
exists; this effectively handles the fact that the MSVC STL does
not currently respect the `ERROR_BAD_PATHNAME` Win32 error code.